Reference

LovaszTheta.TH — Function

TH(g::AbstractGraph)

Creates the theta body of the supplied graph. The * operator can be used to scale this theta body by a non-negative constant or by a scalar Convex.Variable.

julia> problem = maximize(sum(w), [w in TH(g)])

julia> problem = minimize(λ, [w in λ*TH(g)])

LovaszTheta.theta_dual_weight — Method

theta_dual_weight(g::AbstractGraph, w::AbstractArray{<:Number, 1})

Compute the weight v that saturates the inequality θ(g, w) * θ(complement(g), v) ≥ w' v.

The returned weight vector will be normalized such that θ(complement(g), v) = 1.

LovaszTheta.theta_solution — Method

theta_solution(g::AbstractGraph, w::AbstractArray{<:Number, 1})

Compute the optimal matrices for the various formulations of θ. Returns a named tuple (λ, R, Y, Z, B, T) satisfying the following:

λ = θ(g, w)

R[i,j] = sqrt(w[i]*w[j])

Y ⪰ R
diag(Y) = λ*I
Y[i,j] = 0 unless i=j or i∼j

diag(Z) = w
[λ w'; w Z] ⪰ 0
Z[i,j] = 0 unless i=j or i∼j

B ⪰ 0
tr(B) = 1
tr(B*R) = λ
B[i,j] = 0 for i∼j

T ⪰ 0
diag(T) = w
|T| = λ
T[i,j] = 0 for i∼j

LovaszTheta.θ — Method

θ(g::AbstractGraph, w::AbstractArray{<:Number, 2})

Compute the matrix weighted Lovasz theta function for the given graph.

LovaszTheta.θ — Method

θ(g::AbstractGraph, w::AbstractArray{<:Number, 1})

Compute the weighted Lovasz theta function for the given graph.

LovaszTheta.θ — Method

θ(g::AbstractGraph, w::Convex.AbstractExpr)

Returns λ bounded from below by θ(g, w). The constraint θ(g, w) ≤ 1 is equivalent to w ∈ TH(complement(g)), the theta body of the complement graph.

LovaszTheta.θ — Method

θ(g::AbstractGraph)

Compute the Lovasz theta function for the given graph.

LovaszTheta.θ⁺ — Method

θ(g::AbstractGraph, w::AbstractArray{<:Number, 1})

Compute the weighted Szegedy theta function for the given graph.

LovaszTheta.θ⁺ — Method

θ(g::AbstractGraph)

Compute the Szegedy theta function for the given graph.

LovaszTheta.θ⁻ — Method

θ(g::AbstractGraph, w::AbstractArray{<:Number, 1})

Compute the weighted Schrijver theta function for the given graph.

LovaszTheta.θ⁻ — Method

θ(g::AbstractGraph)

Compute the Schrijver theta function for the given graph.